// Be-Instance/src/types.ts
export type MasterTier = 'sys_arch' | 'user' | 'immersive';

export type SubViewMode = 
  | 'millennium_rumination'
  | 'synesthetic_manifold'
  | 'framebuffer_mirror'
  | 'canvas_4d_projection'
  | 'organelle_synthesis'
  | 'substrate_thermo'
  | 'transpile_forge'
  | 'baremetal_ring0'
  | 'asterion_mediation'
  | 'covalent_asm'
  | 'predicate_metrics_asm'
  | 'singleton_publisher'
  | 'be_repo_viewer'
  | 'epistemic_vault'
  | 'fleet_matrix'
  | 'immune_sandbox'
  | 'mesh_socket_ipc'
  | 'dyad_crucible'
  | 'multimodal_cointeractive'
  | 'dissertation_20ch'
  | 'quipu_memory'
  | 'quotient_sieve'
  | 'zumas_kinetic'
  | 'twin_playpark'
  | 'midi_432hz'
  | 'hybrid_dual_face';